Update block information in UI at least every 5 seconds
During IBD we can get into a long loop in ActivateBestChainStep
and blocks on less powerful system can take long to validate. As
a result in the  UI the user may think nothing is happening as it
takes quite long for the UI to update.  So here we ensure that
at least every 5 seconds the UI will update if a new block has
been processed rather than waiting until all blocks in this loop
were finally connected.

We use a static atomic int because there may be other threads
running concurrently due to Parallel Validation.