Check for any repeating transaction id sequences
This was previously being done when we first received the xthin
however we did not necessarily have all the hashes at that point.

Also, by putting the check in ReconstructBlock() we ensure that
every block reconstruction is checked for repeating sequences
regardless of whether it is from an xthinblock and thinblock or
a reconstruction from an xblocktx.