Make it so that uacomment[] are correctly reported in subver string.
This is a fix for issue #1601. Namely "uaccoments" parameters were
overwritten by EB/AD settings. These change are always appending
"uaccoments" to subver string in the init step and also at run time
when EB and/or AD are changed.

If the resulting subver is longer than 256 chars (per BIP 014), in the
init step we are going to emit a warning and exit. If this situation triggers
at run time we just trim the subver until it met the 256 chars limit.