Remove template matching and pseudo opcodes
The current code contains a rather complex script template matching engine,
which is only used for 3 particular script types (P2PK, P2PKH, multisig).
The first two of these are trivial to match for otherwise, and a specialized
matcher for multisig is both more compact and more efficient than a generic
one.

The goal is being more flexible, so that for example larger standard multisigs
inside SegWit outputs are more easy to implement.

As a side-effect, it also gets rid of the pseudo opcodes hack.

Committer's note:

This is a port of bitcoin/bitcoin/pull/13194 tha has been adapted to
work with BU. The main differencies are related to the fact that we
expose to the user two other kind of transaction use cases: TX_LABELPUBLIC, TX_CLTV.